refactor: 1. folder structure 2. variable name
