+++ src/custom_actions/update_branch.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 
-from utils.command_helper import fire_commands, fire_commands_respectively
+from utils.command import fire_commands, fire_commands_respectively
 from models import config
 from models.status import GitRepoStatus
